refactor: move gemini sdk to the background
