+++ apps/ai-assistant/package.json
@@ -10,6 +10,7 @@
     "@douyinfe/semi-icons": "^2.51.3",
     "@douyinfe/semi-ui": "^2.51.3",
     "@google/generative-ai": "^0.1.3",
+    "@webx-kit/messaging": "workspace:^",
     "@webx-kit/runtime": "workspace:^",
     "@webx-kit/storage": "workspace:^",
     "clsx": "^2.1.0",
