+++ apps/ai-assistant/src/background/index.ts
@@ -1,1 +1,1 @@
-console.log('hello');
+import './router';
